% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_pacientes}
\alias{base_pacientes}
\title{Base de datos de pacientes}
\format{
Un data frame con X filas y Y variables:
\describe{
\item{id}{Identificador único del paciente}
\item{nombre}{Nombre del paciente}
\item{edad}{Edad del paciente en años}
\item{diagnostico}{Diagnóstico principal}
\item{fecha_admision}{Fecha de admisión}
...
}
}
\source{
\url{https://www.ejemplo.com/origen-datos} o "Datos recopilados por..."
}
\usage{
base_pacientes
}
\description{
Conjunto de datos que contiene información sobre pacientes y sus diagnósticos
}
\keyword{datasets}
